+++ projects/swimlane/ngx-ui/src/lib/components/calendar/calender.component.spec.ts
@@ -330,15 +330,15 @@ describe('CalendarComponent', () => {
     it('should update hour correctly', () => {
       component.range = { startDate: new Date(), endDate: undefined };
       component.startAmPmVal = 'PM';
-      component.startHour = 5;
-      component.hourChanged(3, 'start');
-      expect(component.range?.startDate?.getHours()).toBe(15);
+      component.startHour = '5';
+      component.hourChanged('03', 'start');
+      expect(component.range?.startDate?.getHours()).toBe(3);
     });
 
     it('should update minute correctly', () => {
       component.range = { startDate: new Date(), endDate: undefined };
-      component.startMinute = 30;
-      component.minuteChanged(45, 'start');
+      component.startMinute = '30';
+      component.minuteChanged('45', 'start');
       expect(component.range?.startDate?.getMinutes()).toBe(45);
     });
 
@@ -359,22 +359,22 @@ describe('CalendarComponent', () => {
       component.range = { startDate: rangeStart, endDate: rangeEnd };
       component.initializeTime();
 
-      expect(component.startHour).toBe(10);
-      expect(component.startMinute).toBe(30);
+      expect(component.startHour).toBe('10');
+      expect(component.startMinute).toBe('30');
       expect(component.startAmPmVal).toBe('AM');
 
-      expect(component.endHour).toBe(3);
-      expect(component.endMinute).toBe(45);
+      expect(component.endHour).toBe('03');
+      expect(component.endMinute).toBe('45');
       expect(component.endAmPmVal).toBe('PM');
     });
 
     it('should initialize time properties with default values if range start and rangeEnd are not provided', () => {
       component.initializeTime();
 
-      expect(component.startHour).toBe(0);
-      expect(component.endHour).toBe(0);
-      expect(component.startMinute).toBe(0);
-      expect(component.endMinute).toBe(0);
+      expect(component.startHour).toBe('12');
+      expect(component.endHour).toBe('12');
+      expect(component.startMinute).toBe('00');
+      expect(component.endMinute).toBe('00');
       expect(component.startAmPmVal).toBe('AM');
       expect(component.endAmPmVal).toBe('AM');
     });
@@ -385,8 +385,8 @@ describe('CalendarComponent', () => {
       const day = { date: moment('2024-04-04'), nextMonth: true, prevMonth: true } as CalendarDay;
       component.focusDate = moment('2024-04-04');
       component.range = { startDate: undefined, endDate: undefined };
-      component.startHour = 10;
-      component.startMinute = 30;
+      component.startHour = '10';
+      component.startMinute = '30';
 
       component.onDaySelectRange(day);
 
@@ -398,8 +398,8 @@ describe('CalendarComponent', () => {
       const day = { date: moment('2024-04-10'), nextMonth: true, prevMonth: true } as CalendarDay;
       component.focusDate = moment('2024-04-10');
       component.range = { startDate: new Date('2024-04-04T10:30:00'), endDate: undefined };
-      component.endHour = 15;
-      component.endMinute = 45;
+      component.endHour = '15';
+      component.endMinute = '45';
 
       component.onDaySelectRange(day);
 
@@ -411,8 +411,8 @@ describe('CalendarComponent', () => {
       const day = { date: moment('2024-04-01'), nextMonth: true, prevMonth: true } as CalendarDay;
       component.focusDate = moment('2024-04-01');
       component.range = { startDate: new Date('2024-04-04T10:30:00'), endDate: undefined };
-      component.startHour = 8;
-      component.startMinute = 15;
+      component.startHour = '8';
+      component.startMinute = '15';
 
       component.onDaySelectRange(day);
 
@@ -424,8 +424,8 @@ describe('CalendarComponent', () => {
       const day = { date: moment('2024-04-04'), nextMonth: true, prevMonth: true } as CalendarDay;
       component.focusDate = moment('2024-04-04');
       component.range = { startDate: new Date('2024-04-01T08:15:00'), endDate: new Date('2024-04-10T15:45:00') };
-      component.startHour = 12;
-      component.startMinute = 0;
+      component.startHour = '12';
+      component.startMinute = '0';
 
       component.onDaySelectRange(day);